refactor message module: process function
