+++ lib/archethic/p2p/message.ex
@@ -4,18 +4,8 @@ defmodule Archethic.P2P.Message do
   """
 
   alias Archethic.{
-    Account,
-    BeaconChain,
-    Contracts,
     Crypto,
-    Election,
-    Mining,
-    P2P,
     P2P.Node,
-    PubSub,
-    Replication,
-    TransactionChain,
-    TaskSupervisor,
     Utils,
     Utils.VarInt
   }
@@ -25,20 +15,15 @@ defmodule Archethic.P2P.Message do
     Summary,
     SummaryAggregate,
     Slot,
-    Subset,
     Slot
   }
-
-  alias Archethic.SelfRepair
 
   alias Archethic.TransactionChain.{
     Transaction,
     Transaction.CrossValidationStamp,
     Transaction.ValidationStamp,
-    TransactionInput,
     TransactionSummary,
     VersionedTransactionInput,
-    Transaction.ValidationStamp.LedgerOperations,
     Transaction.ValidationStamp.LedgerOperations.VersionedUnspentOutput
   }
 
@@ -103,8 +88,6 @@ defmodule Archethic.P2P.Message do
     ReplicatePendingTransactionChain,
     NotifyReplicationValidation
   }
-
-  alias ArchethicWeb.TransactionSubscriber
 
   require Logger
 
@@ -1343,680 +1326,5 @@ defmodule Archethic.P2P.Message do
   Handle a P2P message by processing it and return list of responses to be streamed back to the client
   """
   @spec process(request(), Crypto.key()) :: response()
-  def process(%GetBootstrappingNodes{patch: patch}, _) do
-    top_nodes = P2P.authorized_and_available_nodes()
-
-    closest_nodes =
-      top_nodes
-      |> P2P.nearest_nodes(patch)
-      |> Enum.take(5)
-
-    %BootstrappingNodes{
-      new_seeds: Enum.take_random(top_nodes, 5),
-      closest_nodes: closest_nodes
-    }
-  end
-
-  def process(%GetStorageNonce{public_key: public_key}, _) do
-    %EncryptedStorageNonce{
-      digest: Crypto.encrypt_storage_nonce(public_key)
-    }
-  end
-
-  def process(%ListNodes{}, _) do
-    %NodeList{nodes: P2P.list_nodes()}
-  end
-
-  def process(%NewTransaction{transaction: tx}, sender_public_key) do
-    case Archethic.send_new_transaction(tx, sender_public_key) do
-      :ok ->
-        %Ok{}
-
-      {:error, :network_issue} ->
-        %Error{reason: :network_issue}
-    end
-  end
-
-  def process(%ValidationError{context: context, reason: reason, address: address}, _) do
-    TransactionSubscriber.report_error(address, context, reason)
-    %Ok{}
-  end
-
-  def process(%GetTransaction{address: tx_address}, _) do
-    case TransactionChain.get_transaction(tx_address) do
-      {:ok, tx} ->
-        tx
-
-      {:error, :transaction_not_exists} ->
-        %NotFound{}
-
-      {:error, :invalid_transaction} ->
-        %Error{reason: :invalid_transaction}
-    end
-  end
-
-  # paging_state recieved  contains binary offset for next page , to be used for query
-  def process(
-        %GetTransactionChain{
-          address: tx_address,
-          paging_state: paging_state,
-          order: order
-        },
-        _
-      ) do
-    {chain, more?, paging_state} =
-      tx_address
-      |> TransactionChain.get([], paging_state: paging_state, order: order)
-
-    # empty list for fields/cols to be processed
-    # new_page_state contains binary offset for the next page
-    %TransactionList{transactions: chain, paging_state: paging_state, more?: more?}
-  end
-
-  def process(%GetUnspentOutputs{address: tx_address, offset: offset}, _) do
-    utxos = Account.get_unspent_outputs(tx_address)
-    utxos_length = length(utxos)
-
-    %{utxos: utxos, offset: offset, more?: more?} =
-      utxos
-      # |> Enum.sort_by(& &1.timestamp, {:desc, DateTime})
-      |> Enum.with_index()
-      |> Enum.drop(offset)
-      |> Enum.reduce_while(%{utxos: [], offset: 0, more?: false}, fn {versioned_utxo, index},
-                                                                     acc ->
-        acc_size =
-          acc.utxos
-          |> Enum.map(&VersionedUnspentOutput.serialize/1)
-          |> :erlang.list_to_binary()
-          |> byte_size()
-
-        utxo_size =
-          versioned_utxo
-          |> VersionedUnspentOutput.serialize()
-          |> byte_size
-
-        if acc_size + utxo_size < 3_000_000 do
-          new_acc =
-            acc
-            |> Map.update!(:utxos, &[versioned_utxo | &1])
-            |> Map.put(:offset, index + 1)
-            |> Map.put(:more?, index + 1 < utxos_length)
-
-          {:cont, new_acc}
-        else
-          {:halt, acc}
-        end
-      end)
-
-    %UnspentOutputList{
-      unspent_outputs: utxos,
-      offset: offset,
-      more?: more?
-    }
-  end
-
-  def process(%GetP2PView{node_public_keys: node_public_keys}, _) do
-    nodes =
-      Enum.map(node_public_keys, fn key ->
-        {:ok, node} = P2P.get_node_info(key)
-        node
-      end)
-
-    view = P2P.nodes_availability_as_bits(nodes)
-    %P2PView{nodes_view: view}
-  end
-
-  def process(
-        %StartMining{
-          transaction: tx = %Transaction{},
-          welcome_node_public_key: welcome_node_public_key,
-          validation_node_public_keys: validation_nodes
-        },
-        _
-      ) do
-    with {:election, true} <- {:election, Mining.valid_election?(tx, validation_nodes)},
-         {:elected, true} <-
-           {:elected, Enum.any?(validation_nodes, &(&1 == Crypto.last_node_public_key()))},
-         {:mining, false} <- {:mining, Mining.processing?(tx.address)} do
-      {:ok, _} = Mining.start(tx, welcome_node_public_key, validation_nodes)
-      %Ok{}
-    else
-      {:election, false} ->
-        Logger.error("Invalid validation node election",
-          transaction_address: Base.encode16(tx.address),
-          transaction_type: tx.type
-        )
-
-        %Error{reason: :network_issue}
-
-      {:elected, false} ->
-        Logger.error("Unexpected start mining message",
-          transaction_address: Base.encode16(tx.address),
-          transaction_type: tx.type
-        )
-
-        %Error{reason: :network_issue}
-
-      {:mining, true} ->
-        Logger.warning("Transaction already in mining process",
-          transaction_address: Base.encode16(tx.address),
-          transaction_type: tx.type
-        )
-
-        %Ok{}
-    end
-  end
-
-  def process(
-        %AddMiningContext{
-          address: tx_address,
-          validation_node_public_key: validation_node,
-          previous_storage_nodes_public_keys: previous_storage_nodes_public_keys,
-          chain_storage_nodes_view: chain_storage_nodes_view,
-          beacon_storage_nodes_view: beacon_storage_nodes_view,
-          io_storage_nodes_view: io_storage_nodes_view
-        },
-        _
-      ) do
-    :ok =
-      Mining.add_mining_context(
-        tx_address,
-        validation_node,
-        previous_storage_nodes_public_keys,
-        chain_storage_nodes_view,
-        beacon_storage_nodes_view,
-        io_storage_nodes_view
-      )
-
-    %Ok{}
-  end
-
-  def process(
-        %ReplicateTransactionChain{
-          transaction:
-            tx = %Transaction{
-              address: tx_address,
-              type: tx_type,
-              validation_stamp: %ValidationStamp{timestamp: timestamp}
-            },
-          replying_node: replying_node_public_key
-        },
-        _
-      ) do
-    # We don't check the election for network transactions because all the nodes receive the chain replication message
-    # The chain storage nodes election is all the authorized nodes but during I/O replication, we send this message to enforce
-    # the synchronization of the network chains
-    if Transaction.network_type?(tx_type) do
-      process_replication_chain(tx, replying_node_public_key)
-    else
-      storage_nodes =
-        Election.chain_storage_nodes_with_type(
-          tx_address,
-          tx_type,
-          P2P.authorized_and_available_nodes(timestamp)
-        )
-
-      # Replicate transaction chain only if the current node is one of the chain storage nodes
-      if Utils.key_in_node_list?(storage_nodes, Crypto.first_node_public_key()) do
-        process_replication_chain(tx, replying_node_public_key)
-      end
-    end
-
-    %Ok{}
-  end
-
-  def process(
-        %ReplicateTransaction{
-          transaction:
-            tx = %Transaction{validation_stamp: %ValidationStamp{timestamp: validation_time}}
-        },
-        _
-      ) do
-    resolved_addresses = TransactionChain.resolve_transaction_addresses(tx, validation_time)
-
-    io_storage_nodes =
-      if Transaction.network_type?(tx.type) do
-        P2P.list_nodes()
-      else
-        resolved_addresses
-        |> Enum.map(fn {_origin, resolved} -> resolved end)
-        |> Enum.concat([LedgerOperations.burning_address()])
-        |> Election.io_storage_nodes(P2P.authorized_and_available_nodes(validation_time))
-      end
-
-    # Replicate tx only if the current node is one of the I/O storage nodes
-    if Utils.key_in_node_list?(io_storage_nodes, Crypto.first_node_public_key()) do
-      case Replication.validate_and_store_transaction(tx) do
-        :ok ->
-          %Ok{}
-
-        {:error, reason} ->
-          %ReplicationError{address: tx.address, reason: reason}
-      end
-    else
-      %Ok{}
-    end
-  end
-
-  def process(
-        %AcknowledgeStorage{
-          address: address,
-          signature: signature
-        },
-        node_public_key
-      ) do
-    Mining.confirm_replication(address, signature, node_public_key)
-    %Ok{}
-  end
-
-  def process(
-        %ReplicationError{
-          address: address,
-          reason: reason
-        },
-        from
-      ) do
-    Mining.notify_replication_error(address, reason, from)
-    %Ok{}
-  end
-
-  def process(
-        %CrossValidate{
-          address: tx_address,
-          validation_stamp: stamp,
-          replication_tree: replication_tree,
-          confirmed_validation_nodes: confirmed_validation_nodes
-        },
-        _
-      ) do
-    Mining.cross_validate(tx_address, stamp, replication_tree, confirmed_validation_nodes)
-    %Ok{}
-  end
-
-  def process(%CrossValidationDone{address: tx_address, cross_validation_stamp: stamp}, _) do
-    Mining.add_cross_validation_stamp(tx_address, stamp)
-    %Ok{}
-  end
-
-  def process(%NotifyEndOfNodeSync{node_public_key: public_key, timestamp: timestamp}, _) do
-    BeaconChain.add_end_of_node_sync(public_key, timestamp)
-    %Ok{}
-  end
-
-  def process(%GetLastTransaction{address: address}, _) do
-    case TransactionChain.get_last_transaction(address) do
-      {:ok, tx} ->
-        tx
-
-      {:error, :transaction_not_exists} ->
-        %NotFound{}
-
-      {:error, :invalid_transaction} ->
-        %Error{reason: :invalid_transaction}
-    end
-  end
-
-  def process(%GetBalance{address: address}, _) do
-    %{uco: uco, token: token} = Account.get_balance(address)
-
-    %Balance{
-      uco: uco,
-      token: token
-    }
-  end
-
-  def process(%GetTransactionInputs{address: address, offset: offset, limit: limit}, _) do
-    contract_inputs =
-      address
-      |> Contracts.list_contract_transactions()
-      |> Enum.map(fn {address, timestamp, protocol_version} ->
-        %VersionedTransactionInput{
-          input: %TransactionInput{from: address, type: :call, timestamp: timestamp},
-          protocol_version: protocol_version
-        }
-      end)
-
-    inputs = Account.get_inputs(address) ++ contract_inputs
-    inputs_length = length(inputs)
-
-    %{inputs: inputs, offset: offset, more?: more?} =
-      inputs
-      |> Enum.sort_by(& &1.input.timestamp, {:desc, DateTime})
-      |> Enum.with_index()
-      |> Enum.drop(offset)
-      |> Enum.reduce_while(
-        %{inputs: [], offset: 0, more?: false},
-        fn {versioned_input, index}, acc = %{inputs: inputs, offset: offset, more?: more?} ->
-          acc_size = Map.get(acc, :acc_size, 0)
-          acc_length = Map.get(acc, :acc_length, 0)
-
-          input_size =
-            versioned_input
-            |> VersionedTransactionInput.serialize()
-            |> byte_size
-
-          size_capacity? = acc_size + input_size < 3_000_000
-
-          should_take_more? =
-            if limit > 0 do
-              acc_length < limit and size_capacity?
-            else
-              size_capacity?
-            end
-
-          if should_take_more? do
-            new_acc =
-              acc
-              |> Map.update!(:inputs, &[versioned_input | &1])
-              |> Map.update(:acc_size, input_size, &(&1 + input_size))
-              |> Map.update(:acc_length, 1, &(&1 + 1))
-              |> Map.put(:offset, index + 1)
-              |> Map.put(:more?, index + 1 < inputs_length)
-
-            {:cont, new_acc}
-          else
-            {:halt, %{inputs: inputs, offset: offset, more?: more?}}
-          end
-        end
-      )
-
-    %TransactionInputList{
-      inputs: Enum.reverse(inputs),
-      more?: more?,
-      offset: offset
-    }
-  end
-
-  # Returns the length of the transaction chain
-  def process(%GetTransactionChainLength{address: address}, _) do
-    %TransactionChainLength{
-      length: TransactionChain.size(address)
-    }
-  end
-
-  # Returns the first public_key for a given public_key and if the public_key is used for the first time, return the same public_key.
-  def process(%GetFirstPublicKey{public_key: public_key}, _) do
-    %FirstPublicKey{
-      public_key: TransactionChain.get_first_public_key(public_key)
-    }
-  end
-
-  def process(%GetGenesisAddress{address: address}, _) do
-    genesis_address = TransactionChain.get_genesis_address(address)
-    %GenesisAddress{address: genesis_address}
-  end
-
-  def process(%GetLastTransactionAddress{address: address, timestamp: timestamp}, _) do
-    {address, time} = TransactionChain.get_last_address(address, timestamp)
-    %LastTransactionAddress{address: address, timestamp: time}
-  end
-
-  def process(
-        %NotifyLastTransactionAddress{
-          genesis_address: genesis_address,
-          last_address: last_address,
-          previous_address: previous_address,
-          timestamp: timestamp
-        },
-        _
-      ) do
-    with {local_last_address, _} <- TransactionChain.get_last_address(genesis_address),
-         true <- local_last_address != last_address do
-      if local_last_address == previous_address do
-        TransactionChain.register_last_address(genesis_address, last_address, timestamp)
-
-        # Stop potential previous smart contract
-        Contracts.stop_contract(local_last_address)
-      else
-        authorized_nodes = P2P.authorized_and_available_nodes()
-        SelfRepair.update_last_address(local_last_address, authorized_nodes)
-      end
-    end
-
-    %Ok{}
-  end
-
-  def process(%GetTransactionSummary{address: address}, _) do
-    case TransactionChain.get_transaction_summary(address) do
-      {:ok, summary} ->
-        summary
-
-      {:error, :not_found} ->
-        %NotFound{}
-    end
-  end
-
-  def process(%GetCurrentSummaries{subsets: subsets}, _) do
-    transaction_summaries =
-      Enum.flat_map(subsets, fn subset ->
-        transaction_summaries = BeaconChain.get_summary_slots(subset)
-
-        %Slot{transaction_attestations: transaction_attestations} =
-          Subset.get_current_slot(subset)
-
-        Enum.reduce(
-          transaction_attestations,
-          transaction_summaries,
-          &[&1.transaction_summary | &2]
-        )
-      end)
-
-    %TransactionSummaryList{
-      transaction_summaries: transaction_summaries
-    }
-  end
-
-  def process(%Ping{}, _), do: %Ok{}
-
-  def process(%GetBeaconSummary{address: address}, _) do
-    case BeaconChain.get_summary(address) do
-      {:ok, summary} ->
-        summary
-
-      {:error, :not_found} ->
-        %NotFound{}
-    end
-  end
-
-  def process(%NewBeaconSlot{slot: slot = %Slot{subset: subset, slot_time: slot_time}}, _) do
-    summary_time = BeaconChain.next_summary_date(slot_time)
-    node_list = P2P.authorized_and_available_nodes(summary_time, true)
-
-    beacon_summary_nodes =
-      Election.beacon_storage_nodes(
-        subset,
-        summary_time,
-        node_list,
-        Election.get_storage_constraints()
-      )
-
-    # Load BeaconChain's slot only for the summary nodes
-    with true <- Utils.key_in_node_list?(beacon_summary_nodes, Crypto.first_node_public_key()),
-         :ok <- BeaconChain.load_slot(slot) do
-      %Ok{}
-    else
-      false ->
-        Logger.error("Unexpected beacon slot broadcast")
-        %Error{reason: :network_issue}
-
-      :error ->
-        %Error{reason: :invalid_transaction}
-    end
-  end
-
-  def process(%GetBeaconSummaries{addresses: addresses}, _) do
-    %BeaconSummaryList{
-      summaries: BeaconChain.get_beacon_summaries(addresses)
-    }
-  end
-
-  def process(%RegisterBeaconUpdates{node_public_key: node_public_key, subset: subset}, _) do
-    BeaconChain.subscribe_for_beacon_updates(subset, node_public_key)
-    %Ok{}
-  end
-
-  def process(%BeaconUpdate{transaction_attestations: transaction_attestations}, sender) do
-    Enum.each(transaction_attestations, fn %ReplicationAttestation{
-                                             transaction_summary: tx_summary
-                                           } ->
-      process(tx_summary, sender)
-    end)
-
-    %Ok{}
-  end
-
-  def process(tx_summary = %TransactionSummary{}, _) do
-    PubSub.notify_transaction_attestation(tx_summary)
-
-    %Ok{}
-  end
-
-  def process(
-        attestation = %ReplicationAttestation{
-          transaction_summary: %TransactionSummary{address: tx_address, type: tx_type}
-        },
-        _
-      ) do
-    case ReplicationAttestation.validate(attestation) do
-      :ok ->
-        PubSub.notify_replication_attestation(attestation)
-        %Ok{}
-
-      {:error, :invalid_confirmations_signatures} ->
-        Logger.error("Invalid attestation signatures",
-          transaction_address: Base.encode16(tx_address),
-          transaction_type: tx_type
-        )
-
-        %Error{reason: :invalid_attestation}
-    end
-  end
-
-  def process(%GetBeaconSummariesAggregate{date: date}, _) do
-    case BeaconChain.get_summaries_aggregate(date) do
-      {:ok, aggregate} ->
-        aggregate
-
-      {:error, :not_exists} ->
-        %NotFound{}
-    end
-  end
-
-  def process(%NotifyPreviousChain{address: address}, _) do
-    Replication.acknowledge_previous_storage_nodes(address)
-    %Ok{}
-  end
-
-  def process(
-        %ShardRepair{
-          first_address: first_address,
-          storage_address: storage_address,
-          io_addresses: io_addresses
-        },
-        _
-      ) do
-    # Ensure all addresses are expected to be replicated
-    nodes = P2P.authorized_and_available_nodes()
-
-    addresses =
-      if storage_address != nil, do: [storage_address | io_addresses], else: io_addresses
-
-    public_key = Crypto.first_node_public_key()
-
-    if Enum.all?(
-         addresses,
-         &(Election.storage_nodes(&1, nodes) |> Utils.key_in_node_list?(public_key))
-       ) do
-      case SelfRepair.repair_in_progress?(first_address) do
-        false ->
-          SelfRepair.start_worker(
-            first_address: first_address,
-            storage_address: storage_address,
-            io_addresses: io_addresses
-          )
-
-        pid ->
-          SelfRepair.add_repair_addresses(pid, storage_address, io_addresses)
-      end
-    end
-
-    %Ok{}
-  end
-
-  def process(%GetNextAddresses{address: address}, _) do
-    case TransactionChain.get_transaction(address, validation_stamp: [:timestamp]) do
-      {:ok, %Transaction{validation_stamp: %ValidationStamp{timestamp: address_timestamp}}} ->
-        addresses =
-          TransactionChain.get_genesis_address(address)
-          |> TransactionChain.list_chain_addresses()
-          |> Enum.filter(fn {_address, timestamp} ->
-            DateTime.compare(timestamp, address_timestamp) == :gt
-          end)
-
-        %AddressList{addresses: addresses}
-
-      _ ->
-        %AddressList{addresses: []}
-    end
-  end
-
-  def process(%ValidateTransaction{transaction: tx}, _) do
-    case Replication.validate_transaction(tx) do
-      :ok ->
-        Replication.add_transaction_to_commit_pool(tx)
-        %Ok{}
-
-      {:error, reason} ->
-        %ReplicationError{address: tx.address, reason: reason}
-    end
-  end
-
-  def process(%ReplicatePendingTransactionChain{address: address}, sender_public_key) do
-    case Replication.get_transaction_in_commit_pool(address) do
-      {:ok, tx} ->
-        Task.Supervisor.start_child(TaskSupervisor, fn ->
-          Replication.sync_transaction_chain(tx)
-          tx_summary = TransactionSummary.from_transaction(tx)
-
-          ack = %AcknowledgeStorage{
-            address: tx.address,
-            signature: Crypto.sign_with_first_node_key(TransactionSummary.serialize(tx_summary))
-          }
-
-          P2P.send_message(sender_public_key, ack)
-        end)
-
-        %Ok{}
-
-      {:error, :transaction_not_exists} ->
-        %Error{reason: :invalid_transaction}
-    end
-  end
-
-  def process(%NotifyReplicationValidation{address: address}, node_public_key) do
-    Mining.notify_replication_validation(address, node_public_key)
-    %Ok{}
-  end
-
-  defp process_replication_chain(tx, replying_node_public_key) do
-    Task.Supervisor.start_child(TaskSupervisor, fn ->
-      response =
-        case Replication.validate_and_store_transaction_chain(tx) do
-          :ok ->
-            tx_summary = TransactionSummary.from_transaction(tx)
-
-            %AcknowledgeStorage{
-              address: tx.address,
-              signature: Crypto.sign_with_first_node_key(TransactionSummary.serialize(tx_summary))
-            }
-
-          {:error, reason} ->
-            %ReplicationError{address: tx.address, reason: reason}
-        end
-
-      if replying_node_public_key do
-        P2P.send_message(replying_node_public_key, response)
-      end
-    end)
-  end
+  def process(msg, key), do: msg.__struct__.process(msg, key)
 end
