+++ lib/archethic/p2p/message/acknowledge_storage.ex
@@ -4,11 +4,29 @@ defmodule Archethic.P2P.Message.AcknowledgeStorage do
 
   This message is used during the transaction replication
   """
-  @enforce_keys [:address, :signature]
-  defstruct [:address, :signature]
+
+  @enforce_keys [:address, :signature, :node_public_key]
+  defstruct [:address, :signature, :node_public_key]
+
+  alias Archethic.Crypto
+  alias Archethic.Mining
+  alias Archethic.P2P.Message.Ok
 
   @type t :: %__MODULE__{
           address: binary(),
           signature: binary()
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: Ok.t()
+  def process(
+        %__MODULE__{
+          address: address,
+          signature: signature,
+          node_public_key: node_public_key
+        },
+        _
+      ) do
+    Mining.confirm_replication(address, signature, node_public_key)
+    %Ok{}
+  end
 end
