+++ lib/archethic/p2p/message/add_mining_context.ex
@@ -21,6 +21,8 @@ defmodule Archethic.P2P.Message.AddMiningContext do
   ]
 
   alias Archethic.Crypto
+  alias Archethic.Mining
+  alias Archethic.P2P.Message.Ok
 
   @type t :: %__MODULE__{
           address: Crypto.versioned_hash(),
@@ -30,4 +32,29 @@ defmodule Archethic.P2P.Message.AddMiningContext do
           io_storage_nodes_view: bitstring(),
           previous_storage_nodes_public_keys: list(Crypto.key())
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: Ok.t()
+  def process(
+        %__MODULE__{
+          address: tx_address,
+          validation_node_public_key: validation_node,
+          previous_storage_nodes_public_keys: previous_storage_nodes_public_keys,
+          chain_storage_nodes_view: chain_storage_nodes_view,
+          beacon_storage_nodes_view: beacon_storage_nodes_view,
+          io_storage_nodes_view: io_storage_nodes_view
+        },
+        _
+      ) do
+    :ok =
+      Mining.add_mining_context(
+        tx_address,
+        validation_node,
+        previous_storage_nodes_public_keys,
+        chain_storage_nodes_view,
+        beacon_storage_nodes_view,
+        io_storage_nodes_view
+      )
+
+    %Ok{}
+  end
 end
