+++ lib/archethic/p2p/message/cross_validation_done.ex
@@ -9,9 +9,17 @@ defmodule Archethic.P2P.Message.CrossValidationDone do
 
   alias Archethic.Crypto
   alias Archethic.TransactionChain.Transaction.CrossValidationStamp
+  alias Archethic.Mining
+  alias Archethic.P2P.Message.Ok
 
   @type t :: %__MODULE__{
           address: Crypto.versioned_hash(),
           cross_validation_stamp: CrossValidationStamp.t()
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: Ok.t()
+  def process(%__MODULE__{address: tx_address, cross_validation_stamp: stamp}, _) do
+    Mining.add_cross_validation_stamp(tx_address, stamp)
+    %Ok{}
+  end
 end
