+++ lib/archethic/p2p/message/get_balance.ex
@@ -6,8 +6,20 @@ defmodule Archethic.P2P.Message.GetBalance do
   defstruct [:address]
 
   alias Archethic.Crypto
+  alias Archethic.Account
+  alias Archethic.P2P.Message.Balance
 
   @type t :: %__MODULE__{
           address: Crypto.versioned_hash()
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: Balance.t()
+  def process(%__MODULE__{address: address}, _) do
+    %{uco: uco, token: token} = Account.get_balance(address)
+
+    %Balance{
+      uco: uco,
+      token: token
+    }
+  end
 end
