+++ lib/archethic/p2p/message/get_beacon_summaries.ex
@@ -4,7 +4,18 @@ defmodule Archethic.P2P.Message.GetBeaconSummaries do
   """
   defstruct [:addresses]
 
+  alias Archethic.Crypto
+  alias Archethic.BeaconChain
+  alias Archethic.P2P.Message.BeaconSummaryList
+
   @type t() :: %__MODULE__{
           addresses: list(binary())
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: BeaconSummaryList.t()
+  def process(%__MODULE__{addresses: addresses}, _) do
+    %BeaconSummaryList{
+      summaries: BeaconChain.get_beacon_summaries(addresses)
+    }
+  end
 end
