+++ lib/archethic/p2p/message/get_beacon_summary.ex
@@ -7,8 +7,22 @@ defmodule Archethic.P2P.Message.GetBeaconSummary do
   defstruct [:address]
 
   alias Archethic.Crypto
+  alias Archethic.BeaconChain
+  alias Archethic.P2P.Message.NotFound
+  alias Archethic.BeaconChain.Summary
 
   @type t :: %__MODULE__{
           address: Crypto.versioned_hash()
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: Summary.t() | NotFound.t()
+  def process(%__MODULE__{address: address}, _) do
+    case BeaconChain.get_summary(address) do
+      {:ok, summary} ->
+        summary
+
+      {:error, :not_found} ->
+        %NotFound{}
+    end
+  end
 end
