+++ lib/archethic/p2p/message/get_genesis_address.ex
@@ -6,7 +6,17 @@ defmodule Archethic.P2P.Message.GetGenesisAddress do
   @enforce_keys [:address]
   defstruct [:address]
 
+  alias Archethic.Crypto
+  alias Archethic.TransactionChain
+  alias Archethic.P2P.Message.FirstAddress
+
   @type t() :: %__MODULE__{
           address: binary()
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: FirstAddress.t()
+  def process(%__MODULE__{address: address}, _) do
+    genesis_address = TransactionChain.get_genesis_address(address)
+    %FirstAddress{address: genesis_address}
+  end
 end
