+++ lib/archethic/p2p/message/get_last_transaction_address.ex
@@ -6,9 +6,17 @@ defmodule Archethic.P2P.Message.GetLastTransactionAddress do
   defstruct [:address, :timestamp]
 
   alias Archethic.Crypto
+  alias Archethic.TransactionChain
+  alias Archethic.P2P.Message.LastTransactionAddress
 
   @type t :: %__MODULE__{
           address: Crypto.versioned_hash(),
           timestamp: DateTime.t()
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: LastTransactionAddress.t()
+  def process(%__MODULE__{address: address, timestamp: timestamp}, _) do
+    {address, time} = TransactionChain.get_last_address(address, timestamp)
+    %LastTransactionAddress{address: address, timestamp: time}
+  end
 end
