+++ lib/archethic/p2p/message/get_storage_nonce.ex
@@ -4,6 +4,8 @@ defmodule Archethic.P2P.Message.GetStorageNonce do
 
   This message is used during the node bootstrapping
   """
+
+  alias Archethic.P2P.Message.EncryptedStorageNonce
   alias Archethic.Crypto
 
   @enforce_keys [:public_key]
@@ -12,4 +14,11 @@ defmodule Archethic.P2P.Message.GetStorageNonce do
   @type t :: %__MODULE__{
           public_key: Crypto.key()
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: EncryptedStorageNonce.t()
+  def process(%__MODULE__{public_key: public_key}, _) do
+    %EncryptedStorageNonce{
+      digest: Crypto.encrypt_storage_nonce(public_key)
+    }
+  end
 end
