+++ lib/archethic/p2p/message/get_transaction_chain_length.ex
@@ -6,8 +6,18 @@ defmodule Archethic.P2P.Message.GetTransactionChainLength do
   defstruct [:address]
 
   alias Archethic.Crypto
+  alias Archethic.TransactionChain
+  alias Archethic.P2P.Message.TransactionChainLength
 
   @type t :: %__MODULE__{
           address: Crypto.versioned_hash()
         }
+
+  # Returns the length of the transaction chain
+  @spec process(__MODULE__.t(), Crypto.key()) :: TransactionChainLength.t()
+  def process(%__MODULE__{address: address}, _) do
+    %TransactionChainLength{
+      length: TransactionChain.size(address)
+    }
+  end
 end
