+++ lib/archethic/p2p/message/list_nodes.ex
@@ -2,7 +2,16 @@ defmodule Archethic.P2P.Message.ListNodes do
   @moduledoc """
   Represents a message to fetch the list of nodes
   """
+  alias Archethic.Crypto
+  alias Archethic.P2P
+  alias Archethic.P2P.Message.NodeList
+
   defstruct []
 
   @type t :: %__MODULE__{}
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: NodeList.t()
+  def process(%__MODULE__{}, _) do
+    %NodeList{nodes: P2P.list_nodes()}
+  end
 end
