+++ lib/archethic/p2p/message/new_transaction.ex
@@ -7,8 +7,22 @@ defmodule Archethic.P2P.Message.NewTransaction do
   defstruct [:transaction]
 
   alias Archethic.TransactionChain.Transaction
+  alias Archethic.P2P.Message.Ok
+  alias Archethic.P2P.Message.Error
+  alias Archethic.Crypto
 
   @type t :: %__MODULE__{
           transaction: Transaction.t()
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: Ok.t() | Error.t()
+  def process(%__MODULE__{transaction: tx}, sender_public_key) do
+    case Archethic.send_new_transaction(tx, sender_public_key) do
+      :ok ->
+        %Ok{}
+
+      {:error, :network_issue} ->
+        %Error{reason: :network_issue}
+    end
+  end
 end
