+++ lib/archethic/p2p/message/ping.ex
@@ -5,5 +5,11 @@ defmodule Archethic.P2P.Message.Ping do
 
   defstruct []
 
+  alias Archethic.Crypto
+  alias Archethic.P2P.Message.Ok
+
   @type t :: %__MODULE__{}
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: Ok.t()
+  def process(%__MODULE__{}, _), do: %Ok{}
 end
