+++ lib/archethic/p2p/message/register_beacon_updates.ex
@@ -2,7 +2,11 @@ defmodule Archethic.P2P.Message.RegisterBeaconUpdates do
   @moduledoc """
   Represents a message to get a beacon updates
   """
+
   alias Archethic.Crypto
+  alias Archethic.BeaconChain
+  alias Archethic.P2P.Message.Ok
+
   @enforce_keys [:node_public_key, :subset]
   defstruct [:node_public_key, :subset]
 
@@ -10,4 +14,10 @@ defmodule Archethic.P2P.Message.RegisterBeaconUpdates do
           node_public_key: Crypto.key(),
           subset: binary()
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: Ok.t()
+  def process(%__MODULE__{node_public_key: node_public_key, subset: subset}, _) do
+    BeaconChain.subscribe_for_beacon_updates(subset, node_public_key)
+    %Ok{}
+  end
 end
