+++ lib/archethic/p2p/message/replicate_transaction_chain.ex
@@ -6,10 +6,81 @@ defmodule Archethic.P2P.Message.ReplicateTransactionChain do
   defstruct [:transaction, :replying_node]
 
   alias Archethic.Crypto
+  alias Archethic.Election
+  alias Archethic.Replication
+  alias Archethic.P2P
+  alias Archethic.P2P.Message.Ok
+  alias Archethic.Utils
   alias Archethic.TransactionChain.Transaction
+  alias Archethic.TransactionChain.Transaction.ValidationStamp
+  alias Archethic.TransactionChain.TransactionSummary
+  alias Archethic.P2P.Message.AcknowledgeStorage
+  alias Archethic.P2P.Message.ReplicationError
 
   @type t :: %__MODULE__{
           transaction: Transaction.t(),
           replying_node: nil | Crypto.key()
         }
+
+  @spec process(__MODULE__.t(), Crypto.key()) :: Ok.t()
+  def process(
+        %__MODULE__{
+          transaction:
+            tx = %Transaction{
+              address: tx_address,
+              type: tx_type,
+              validation_stamp: %ValidationStamp{timestamp: timestamp}
+            },
+          replying_node: replying_node_public_key
+        },
+        _
+      ) do
+    # We don't check the election for network transactions because all the nodes receive the chain replication message
+    # The chain storage nodes election is all the authorized nodes but during I/O replication, we send this message to enforce
+    # the synchronization of the network chains
+    if Transaction.network_type?(tx_type) do
+      process_replication_chain(tx, replying_node_public_key)
+    else
+      storage_nodes =
+        Election.chain_storage_nodes_with_type(
+          tx_address,
+          tx_type,
+          P2P.authorized_and_available_nodes(timestamp)
+        )
+
+      # Replicate transaction chain only if the current node is one of the chain storage nodes
+      if Utils.key_in_node_list?(storage_nodes, Crypto.first_node_public_key()) do
+        process_replication_chain(tx, replying_node_public_key)
+      end
+    end
+
+    %Ok{}
+  end
+
+  defp process_replication_chain(tx, replying_node_public_key) do
+    Task.Supervisor.start_child(TaskSupervisor, fn ->
+      response =
+        case Replication.validate_and_store_transaction_chain(tx) do
+          :ok ->
+            tx_summary = TransactionSummary.from_transaction(tx)
+
+            %AcknowledgeStorage{
+              address: tx.address,
+              signature:
+                Crypto.sign_with_first_node_key(TransactionSummary.serialize(tx_summary)),
+              node_public_key: Crypto.first_node_public_key()
+            }
+
+          {:error, :transaction_already_exists} ->
+            %ReplicationError{address: tx.address, reason: :transaction_already_exists}
+
+          {:error, invalid_tx_error} ->
+            %ReplicationError{address: tx.address, reason: invalid_tx_error}
+        end
+
+      if replying_node_public_key do
+        P2P.send_message(replying_node_public_key, response)
+      end
+    end)
+  end
 end
